PPU shading code fix so both processors produces same picture

+++ programs/floattype.h
@@ -2,11 +2,11 @@
 #define _H_FLOATTYPE_
 
 #ifdef _PPU_
-typedef float FLOAT_TYPE;
+typedef double FLOAT_TYPE;
 #endif
 
 #ifdef _SPU_
-typedef float FLOAT_TYPE;
+typedef double FLOAT_TYPE;
 #endif
